FINANCE/Trade-Execution-LogAlert/trade-logger/requirements.txt
flask 
boto3 
python-dotenv